+++ Spring/src/main/java/com/leonardo/Spring/resource/SaleResource.java
@@ -2,9 +2,9 @@ package com.leonardo.Spring.resource;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -31,14 +31,23 @@ public class SaleResource {
     // Get All Sale(s)
     @GetMapping
     public ResponseEntity<List<Sale>> getAllSales() {
-        return new ResponseEntity<List<Sale>>(saleService.findAllSales(), HttpStatus.OK);
+
+        List<Sale> sales = saleService.findAllSales();
+
+        if (!sales.isEmpty()) {
+            return new ResponseEntity<List<Sale>>(sales, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<List<Sale>>(HttpStatus.NO_CONTENT);
+        }
     }
 
     // Get Sale by ID
     @GetMapping
     @RequestMapping(value = "/{id}")
     public ResponseEntity<Sale> getSaleById(@PathVariable("id") Long id) {
+
         Sale sale = saleService.findSaleById(id);
+
         if (sale != null) {
             return new ResponseEntity<Sale>(sale, HttpStatus.OK);
         } else {
@@ -48,16 +57,22 @@ public class SaleResource {
 
     // Save Sale
     @PostMapping
-    public ResponseEntity<Sale> addSale(@RequestBody Sale sale) {
-        return new ResponseEntity<Sale>(saleService.saveSale(sale), HttpStatus.CREATED);
+    public ResponseEntity<Void> addSale(@RequestBody @Valid Sale sale) {
+
+        saleService.saveSale(sale);
+
+        return new ResponseEntity<Void>(HttpStatus.CREATED);
     }
 
     // Update Sale
     @PutMapping
     public ResponseEntity<Sale> updateSale(@RequestBody Sale sale) {
+
         Sale oldSale = saleService.findSaleById(sale.getId());
+
         if (oldSale != null) {
-            return new ResponseEntity<Sale>(saleService.saveSale(sale), HttpStatus.OK);
+            saleService.saveSale(sale);
+            return new ResponseEntity<Sale>(sale, HttpStatus.OK);
         } else {
             return new ResponseEntity<Sale>(HttpStatus.NOT_FOUND);
         }
@@ -65,16 +80,15 @@ public class SaleResource {
 
     // Delete Sale
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<String> deleteSale(@PathVariable("id") Long id) {
-        try {
+    public ResponseEntity<Void> deleteSale(@PathVariable("id") Long id) {
+
+        Sale oldSale = saleService.findSaleById(id);
+
+        if (oldSale != null) {
             saleService.deleteSaleById(id);
-            return new ResponseEntity<String>("Sale Successfully Deleted! ID: " + id, HttpStatus.OK);
-        } catch (DataIntegrityViolationException e) {
-            return new ResponseEntity<String>("Failed to Delete Sale Due to Data Integrity Violation! ID: " + id,
-                    HttpStatus.BAD_REQUEST);
-        } catch (EmptyResultDataAccessException e) {
-            return new ResponseEntity<String>("Failed to Delete Sale Due to Not Finding It! ID: " + id,
-                    HttpStatus.NOT_FOUND);
+            return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
         }
     }
 }
